requirements.txt
# Core ML/DL
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# OCR
easyocr>=1.7.0

# Image Processing
Pillow>=9.0.0
numpy>=1.24.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# UI
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0

# PDF Support
pdf2image>=1.16.0
PyMuPDF>=1.23.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Linting
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Type checking
mypy>=1.5.0
